Update documentation and fix minor issues
Update roxygen documentation to include new package functionality. Update vignettes.
Coerce publishWebService and updateWebService to read user defined functions from
the R global environment.

Previously, publishWebService and updateWebService would throw an error if the user passed in
a function that also exists in the packages sourced by AzureML. For examplem, the "add1" function
from the older tutorial would fail to publish as a web service since another function with the
same name exists in the "stats" package. This fix restricts the function lookup to the global
R environment.